Added util for mathematical function
